[assignmentPanel] extract `showOnEncointerMap`
