+++ lib/page-encointer/common/encointerMap.dart
@@ -1,5 +1,6 @@
 import 'package:encointer_wallet/common/theme.dart';
 import 'package:encointer_wallet/store/app.dart';
+import 'package:encointer_wallet/store/encointer/types/location.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -85,4 +86,51 @@ Widget noCommunityDialog(BuildContext context) {
       ),
     ],
   );
+}
+
+Future<void> showOnEncointerMap(
+  BuildContext context,
+  AppStore store,
+  Location location, {
+  double initialZoom = 14,
+}) {
+  var dic = I18n.of(context).translationsForLocale();
+
+  return Navigator.push(
+    context,
+    MaterialPageRoute(
+      builder: (context) {
+        return EncointerMap(
+          store,
+          popupBuilder: (BuildContext context, Marker marker) => SizedBox(),
+          markers: buildMarkers(location),
+          title: dic.encointer.meetupLocation,
+          center: location.toLatLng(),
+          initialZoom: initialZoom,
+        );
+      },
+    ),
+  );
+}
+
+List<Marker> buildMarkers(Location meetupLocation) {
+  List<Marker> markers = <Marker>[];
+  markers.add(
+    Marker(
+      // marker is not a widget, hence test_driver cannot find it (it can find it in the Icon inside, though).
+      // But we need the key to derive the popup key
+      key: Key('meetup-location'),
+      point: meetupLocation.toLatLng(),
+      width: 40,
+      height: 40,
+      builder: (_) => Icon(
+        Icons.location_on,
+        size: 40,
+        color: Colors.blueAccent,
+        key: Key('meetup-location-icon'), // used for test_driver
+      ),
+      anchorPos: AnchorPos.align(AnchorAlign.top),
+    ),
+  );
+  return markers;
 }
